Extract FragmentPainter: paint is its own phase, not a CssBox method
CssBox was still the DOM node, the style store, the layout algorithm and the
painter. Painting now lives in Html/Core/Paint: FragmentPainter consumes a
FragmentainerFragment and owns the whole algorithm - the fragment walk,
transform/clip-path/opacity, tagged-PDF bookkeeping, decorations, text and the
Appendix E child ordering.

Replaced elements no longer override a virtual paint method on the box. Each
kind has an IFragmentContentPainter selected from the box's type, with the four
that share the same shape (image, object, inline SVG, iframe) built on one
ReplacedFragmentPainter base. Boxes expose only the resolved content their
painter reads.

A painter instance owns one page, so the already-painted set moves off
HtmlContainerInt with it. CssBox loses ~1000 lines.

All 60 showcases are byte-identical to the previous build after normalizing
creation dates, document IDs, font subset tags and link-annotation UUIDs; the
only residue is annotation /M timestamps.
