+++ src/PeachPDF/Html/Core/Dom/CssBox.cs
@@ -262,22 +262,13 @@ namespace PeachPDF.Html.Core.Dom
 
         /// <summary>
         /// Set (on the real <c>&lt;body&gt;</c> or <c>&lt;html&gt;</c> box, whichever was chosen) by
-        /// <c>PdfGenerator.ResolveCanvasBackground</c> per CSS2.1 §14.2: that box's background has been
-        /// "promoted" to fill the whole page canvas on every page (see
-        /// <see cref="PaintCanvasBackground"/>), so this box's own normal <see cref="PaintBackground"/>
-        /// call must no-op instead of painting the same background a second time at its own (possibly
+        /// <see cref="HtmlContainerInt.ResolveCanvasBackground"/> per CSS2.1 §14.2: that box's
+        /// background has been "promoted" to fill the whole page canvas on every page (see
+        /// <see cref="FragmentPainter.PaintCanvasBackground"/>), so this box's own normal background
+        /// pass must no-op instead of painting the same background a second time at its own (possibly
         /// much smaller than the page) laid-out rect.
         /// </summary>
         internal bool SuppressOwnBackgroundPaint { get; set; }
-
-        /// <summary>
-        /// Paints this box's own background (see <see cref="PaintBackground"/>) at <paramref name="rect"/>
-        /// instead of this box's own laid-out rect - used by <c>PdfGenerator.AddPdfPages</c> to fill the
-        /// whole page canvas with the <c>&lt;body&gt;</c>/<c>&lt;html&gt;</c> background per CSS2.1 §14.2,
-        /// reusing the exact same background-image/-position/-size/-repeat/-origin/-clip resolution the
-        /// box's own normal paint path uses so canvas-fill behavior matches per-box behavior exactly.
-        /// </summary>
-        internal void PaintCanvasBackground(RGraphics g, RRect rect) => PaintBackground(g, rect, isFirst: true);
 
         /// <summary>
         /// Whether this box declares any background of its own (a visible <c>background-color</c> and/or
@@ -355,7 +346,7 @@ namespace PeachPDF.Html.Core.Dom
         /// <summary>
         /// The vertical line segments (in absolute document coordinates) to draw between adjacent
         /// columns of a multi-column container — one segment per gap per page-row actually used.
-        /// Set by <see cref="CssLayoutEngineColumns"/>, painted by <see cref="PaintImp"/>.
+        /// Set by <see cref="CssLayoutEngineColumns"/>, painted by <see cref="FragmentPainter"/>.
         /// </summary>
         internal List<(double X, double Top, double Bottom)>? ColumnRuleSegments { get; set; }
 
@@ -603,149 +594,6 @@ namespace PeachPDF.Html.Core.Dom
             {
                 HtmlContainer?.ReportError(HtmlRenderErrorType.Layout, "Exception in box layout", ex);
             }
-        }
-
-        /// <summary>
-        /// Paints one of this box's fragments — the portion of it that lives in a single fragmentainer
-        /// (CSS Fragmentation Level 3 §2). All geometry comes from <paramref name="fragment"/>, which
-        /// already carries fragmentainer-local coordinates; this box supplies only style.
-        /// </summary>
-        /// <param name="g">Device context to use</param>
-        /// <param name="fragment">this box's fragment in the fragmentainer being painted</param>
-        internal async ValueTask Paint(RGraphics g, BoxFragment fragment)
-        {
-#if DEBUG
-            Console.WriteLine($"paint: {ToString()}");
-#endif
-
-            try
-            {
-                if (Display == CssConstants.None || Visibility != CssConstants.Visible) return;
-
-                // use initial clip to draw blocks with Position = fixed. I.e. ignore page margins
-                if (Position == CssConstants.Fixed)
-                {
-                    g.SuspendClipping();
-                }
-
-                // A fragment only exists where its box has something in this fragmentainer, so the
-                // cheap early-out is now just "is any of it inside the current clip" - the clip being
-                // narrower than the page band whenever an ancestor's overflow or a clip-path is active.
-                // A fragment that carries only descendants (no decoration rectangles of its own) must
-                // always be entered: a hoisted out-of-flow descendant can paint well outside this
-                // fragment's own rectangle and is only reached through this call.
-                var visible = fragment.Lines.Count == 0 || IsAnyRectVisible(fragment, g.GetClip());
-
-                if (visible)
-                {
-                    var transformed = IsTransformed;
-
-                    if (transformed)
-                    {
-                        // ActualTransformMatrix is cached treating the box's own top-left as local
-                        // (0, 0), so re-anchor the pivot to where the box sits on this page. The whole
-                        // (unfragmented) border box is deliberately used rather than this fragment's
-                        // own rectangle: per-fragment transform origins are a separate change.
-                        g.PushTransform(ActualTransformMatrix.RebaseOrigin(fragment.WholeBoxRect.X, fragment.WholeBoxRect.Y));
-                    }
-
-                    // clip-path clips the entire element rendering (background, border, content, children).
-                    // It is established inside the transform push so the clip and the content it clips are
-                    // transformed together (CSS Masking 1: the clip is in the element's local coordinate
-                    // system, which any `transform` then maps). The reference box is the border-box (the
-                    // default) of the whole box, for the same reason as the transform pivot above.
-                    var clipped = false;
-                    if (ClipPath != CssConstants.None && !string.IsNullOrEmpty(ClipPath))
-                    {
-                        if (CssClipPathResolver.TryBuildClipPath(g, ClipPath, fragment.WholeBoxRect, this, out var clipGeometry, out _)
-                            && clipGeometry is not null)
-                        {
-                            g.PushClip(clipGeometry);
-                            clipped = true;
-                        }
-                    }
-
-                    if (IsOpaque)
-                    {
-                        await PaintImp(g, fragment);
-                    }
-                    else
-                    {
-                        await PaintWithOpacity(g, fragment);
-                    }
-
-                    if (clipped)
-                        g.PopClip();
-
-                    if (transformed)
-                        g.PopTransform();
-                }
-
-                // Restore clips
-                if (Position == CssConstants.Fixed)
-                {
-                    g.ResumeClipping();
-                }
-            }
-            catch (Exception ex)
-            {
-                HtmlContainer?.ReportError(HtmlRenderErrorType.Paint, "Exception in box paint", ex);
-            }
-        }
-
-        /// <summary>
-        /// Whether any of <paramref name="fragment"/>'s own decoration rectangles survives
-        /// <paramref name="clip"/>. Checked per rectangle rather than against their union, so a
-        /// fragment whose rectangles are far apart on the line cannot be kept alive by empty space
-        /// between them.
-        /// </summary>
-        private static bool IsAnyRectVisible(BoxFragment fragment, RRect clip)
-        {
-            foreach (var line in fragment.Lines)
-            {
-                if (IsRectVisible(line.Rect, clip)) return true;
-            }
-
-            return false;
-        }
-
-        /// <summary>
-        /// Paints this box (and, via <see cref="PaintImp"/>, its whole subtree) into an offscreen tile
-        /// sized to the current page's visible clip, then composites that tile onto <paramref name="g"/>
-        /// as a single flattened result at <c>ActualOpacity</c> - the CSS <c>opacity</c> property
-        /// is a group effect (it applies once to the element and everything painted inside it, not to
-        /// each descendant's own paint calls independently), and this is what makes overlapping content
-        /// within the box composite correctly instead of double-blending where it overlaps.
-        /// </summary>
-        /// <remarks>
-        /// The tile is sized to the whole current page-visible rect (not a tight bounding box of this
-        /// box's own content) so that <see cref="PaintImp"/> can keep painting at its normal absolute
-        /// page coordinates unmodified - this box's own possible multiple <see cref="Rectangles"/> (line
-        /// wraps) and any overflowing/absolutely-positioned descendants all land correctly inside the
-        /// tile with no extra translation math, at the cost of a somewhat larger-than-necessary Form
-        /// XObject. Any transform already pushed onto <paramref name="g"/> for this same box (see the
-        /// caller in <see cref="Paint"/>) is left active and applies to the tile's placement automatically,
-        /// since PDF's own <c>cm</c> operator concatenates - no separate transform-folding is needed here.
-        /// </remarks>
-        private async ValueTask PaintWithOpacity(RGraphics g, BoxFragment fragment)
-        {
-            var clip = g.GetClip();
-            var tileRect = new RRect(0, 0, clip.Right, clip.Bottom);
-
-            var tile = g.CreateTile(tileRect.Width, tileRect.Height);
-            if (tile is not { } t)
-            {
-                // No page/document context to own a Form XObject in (e.g. a measure-only pass) -
-                // opacity has no visual effect there anyway, so just paint directly.
-                await PaintImp(g, fragment);
-                return;
-            }
-
-            t.Graphics.PushClip(clip);
-            await PaintImp(t.Graphics, fragment);
-            t.Graphics.Dispose();
-
-            g.DrawImageWithOpacity(t.Image, tileRect, ActualOpacity);
         }
 
         /// <summary>
@@ -1058,8 +906,10 @@ namespace PeachPDF.Html.Core.Dom
         /// own <see cref="CssBoxProperties.ActualFont"/> (or <see cref="CssBoxProperties.ActualSmallCapsFont"/>
         /// for a synthesized small-caps run). <paramref name="styleSource"/> is the box whose font applies -
         /// the owner box, or a <c>::first-line</c> shadow box for a word on the first formatted line.
+        /// Shared by measurement and by <see cref="FragmentPainter"/>, so the two can never disagree
+        /// about which face a word is drawn in.
         /// </summary>
-        private static RFont ResolveWordFont(CssRect word, CssBoxProperties styleSource)
+        internal static RFont ResolveWordFont(CssRect word, CssBoxProperties styleSource)
         {
             if (word.UsesPerCodepointFont && word.Text is { Length: > 0 } text)
             {
@@ -1938,25 +1788,6 @@ namespace PeachPDF.Html.Core.Dom
         protected sealed override CssBoxProperties? GetParent()
         {
             return _parentBox;
-        }
-
-        private void PaintContentImage(RGraphics g, BoxFragment fragment)
-        {
-            if (ContentImage == null) return;
-            foreach (var line in fragment.Lines)
-            {
-                var rect = line.Rect;
-                if (rect.Width <= 0 || rect.Height <= 0) continue;
-                CssImagePainter.Paint(g, ContentImage, layerIndex: 0, isFirst: true,
-                    originRect: rect, clipRect: rect, roundedClipPath: null,
-                    positionList: "0% 0%", sizeList: CssConstants.Auto, repeatList: "no-repeat",
-                    attachmentList: CssConstants.Scroll, viewportRect: rect, box: this,
-                    drawBrush: brush =>
-                    {
-                        g.DrawRectangle(brush, rect.X, rect.Y, rect.Width, rect.Height);
-                        brush.Dispose();
-                    });
-            }
         }
 
         /// <summary>
@@ -2859,859 +2690,6 @@ namespace PeachPDF.Html.Core.Dom
             }
 
             Location = Location with { X = Location.X + amount };
-        }
-
-        /// <summary>
-        /// Paints one fragment, wrapping <see cref="PaintImpCore(RGraphics, BoxFragment)"/> (the actual
-        /// per-box paint logic, overridden by replaced-element subclasses) with tagged-PDF
-        /// structure-tree/marked-content bookkeeping when tagging is enabled
-        /// (<c>HtmlContainer.StructureTagBuilder</c> is non-null). This is the single choke point
-        /// all tagging flows through, mirroring how <see cref="Paint"/> already conditionally wraps
-        /// <see cref="PaintImpCore(RGraphics, BoxFragment)"/>/<see cref="PaintWithOpacity"/> for
-        /// transform/opacity handling. When tagging is disabled this adds one null check and
-        /// otherwise behaves exactly as calling <c>PaintImpCore</c> directly would.
-        /// </summary>
-        /// <param name="g">the device to draw to</param>
-        /// <param name="fragment">the fragment being painted</param>
-        protected async ValueTask PaintImp(RGraphics g, BoxFragment fragment)
-        {
-            // Mirrors PaintImpCore's own early-out: skip classification/tagging entirely for a
-            // fragment that has already been painted on this page.
-            if (HtmlContainer is not null && !HtmlContainer.TryClaimForPaint(fragment))
-                return;
-
-            var builder = HtmlContainer?.StructureTagBuilder;
-            if (builder == null)
-            {
-                await PaintImpCore(g, fragment);
-                return;
-            }
-
-            var classification = StructureTagMapper.Classify(this);
-            switch (classification.Kind)
-            {
-                case StructureTagKind.Artifact:
-                    using (builder.OpenArtifact(g))
-                        await PaintImpCore(g, fragment);
-                    break;
-
-                case StructureTagKind.Grouping when classification.StructureType == Keywords.Li:
-                    await PaintListItem(g, fragment, builder);
-                    break;
-
-                case StructureTagKind.Grouping:
-                    using (builder.OpenGroupingElement(this, classification.StructureType!))
-                        await PaintImpCore(g, fragment);
-                    break;
-
-                case StructureTagKind.Content:
-                    using (builder.OpenContentElement(g, this, classification.StructureType!, classification.AltText))
-                        await PaintImpCore(g, fragment);
-                    break;
-
-                default:
-                    await PaintImpCore(g, fragment);
-                    break;
-            }
-        }
-
-        /// <summary>
-        /// Paints a tagged &lt;li&gt;'s marker and body as separate sibling structure elements under
-        /// "/LI" - "/Lbl" (the marker, via the synthesized <c>::marker</c> child from
-        /// <see cref="IsMarkerPseudoElement"/>, defaulting to "Lbl" per the UA stylesheet's
-        /// <c>li::marker</c> rule, or suppressed if an author set
-        /// <c>li::marker { -peachpdf-pdf-tag-type: none }</c>) and "/LBody" (everything else - not
-        /// itself a separate CSS signal, inferred as "whatever isn't the marker" per the design
-        /// decision this implements). The struct element for "/Lbl" is opened, and its own content
-        /// painted, before "/LBody"'s so the two land in that order under "/LI"'s "/K" (correct
-        /// reading order) - the marker's own on-page paint position is unaffected by this call-order
-        /// swap, since it's driven entirely by pre-computed layout coordinates, not paint order.
-        /// </summary>
-        private async ValueTask PaintListItem(RGraphics g, BoxFragment fragment, StructureTagBuilder builder)
-        {
-            using (builder.OpenGroupingElement(this, Keywords.Li))
-            {
-                var markerFragment = FindMarkerFragment(fragment);
-                if (markerFragment is not null)
-                {
-                    var markerBox = markerFragment.Box;
-                    var markerClassification = StructureTagMapper.Classify(markerBox);
-                    if (markerClassification.Kind == StructureTagKind.None)
-                    {
-                        await markerBox.PaintImpCore(g, markerFragment);
-                    }
-                    else
-                    {
-                        using (builder.OpenContentElement(g, markerBox, markerClassification.StructureType ?? Keywords.Lbl))
-                            await markerBox.PaintImpCore(g, markerFragment);
-                    }
-                }
-
-                using (builder.OpenListItemBodyElement(this))
-                    await PaintImpCore(g, fragment, paintMarkers: false);
-            }
-        }
-
-        /// <summary>
-        /// This fragment's <c>::marker</c> child fragment, if the marker appears on this page at all.
-        /// </summary>
-        private static BoxFragment? FindMarkerFragment(BoxFragment fragment)
-        {
-            foreach (var child in fragment.Children)
-            {
-                if (child.Box.IsMarkerPseudoElement) return child;
-            }
-
-            return null;
-        }
-
-        /// <summary>
-        /// Paints the fragment. Renamed from the historical "PaintImp" - <see cref="PaintImp"/> is
-        /// now a non-virtual wrapper around this method that adds tagged-PDF marked-content/struct-
-        /// element bookkeeping around whatever this override actually paints, when tagging is
-        /// enabled (see StructureTagMapper/StructureTagBuilder). Subclasses needing custom paint
-        /// logic still override this method, exactly as they overrode "PaintImp" before.
-        /// </summary>
-        /// <param name="g">the device to draw to</param>
-        /// <param name="fragment">the fragment being painted, the source of all geometry</param>
-        protected virtual async ValueTask PaintImpCore(RGraphics g, BoxFragment fragment)
-        {
-            await PaintImpCore(g, fragment, paintMarkers: true);
-        }
-
-        /// <summary>
-        /// The actual paint logic, with list-marker painting optionally excluded. Only the base
-        /// <see cref="CssBox"/> needs this overload - it exists so the tagged-PDF &lt;li&gt; path in
-        /// <see cref="PaintImp"/> (always a plain <see cref="CssBox"/>, never one of the replaced-
-        /// element subclasses that override the virtual <see cref="PaintImpCore(RGraphics, BoxFragment)"/>) can
-        /// paint an &lt;li&gt;'s body content and its marker as two separately MCID-tagged, sibling
-        /// structure elements ("/Lbl" and "/LBody") instead of one combined region.
-        /// </summary>
-        private async ValueTask PaintImpCore(RGraphics g, BoxFragment fragment, bool paintMarkers)
-        {
-            if (Display == CssConstants.None ||
-                (Display == CssConstants.TableCell && EmptyCells == CssConstants.Hide && IsSpaceOrEmpty)) return;
-
-            var clipped = RenderUtils.ClipGraphicsByOverflow(g, this, fragment.OriginY);
-
-            // This fragment's own decoration rectangles - one per line box it spans on this page, or a
-            // single whole-border-box rectangle for a block-level box. Already fragmentainer-local.
-            var lines = fragment.Lines;
-            var clip = g.GetClip();
-
-            for (var i = 0; i < lines.Count; i++)
-            {
-                var actualRect = lines[i].Rect;
-
-                if (!IsRectVisible(actualRect, clip)) continue;
-
-                var hasBoxShadow = !string.IsNullOrEmpty(BoxShadow) &&
-                                   !string.Equals(BoxShadow, CssConstants.None, StringComparison.OrdinalIgnoreCase);
-
-                // Outset (drop) shadows paint BEFORE the background so they sit behind the box
-                // (CSS Backgrounds & Borders 3 §5).
-                if (hasBoxShadow)
-                    PaintBoxShadows(g, actualRect, inset: false);
-
-                // A box whose background was "promoted" to fill the whole page canvas (see
-                // PdfGenerator.ResolveCanvasBackground / PaintCanvasBackground) already had it painted
-                // for every page - skip this box's own normal paint pass so it isn't painted twice.
-                if (!SuppressOwnBackgroundPaint)
-                    PaintBackground(g, actualRect, i == 0, GetFirstLineStyleForRect(lines[i].Line));
-
-                // Inset shadows paint AFTER (over) the background, clipped to the padding box.
-                if (hasBoxShadow)
-                    PaintBoxShadows(g, actualRect, inset: true);
-
-                // For multi-page tables, draw the outer bottom border at the page-break Y on
-                // intermediate pages (instead of at actualRect.Bottom which is off-page).
-                // The PerformPaint clip already constrains the side-border top to MarginTop.
-                var rectForBorders = actualRect;
-                if ((Display == CssConstants.Table || Display == CssConstants.InlineTable)
-                    && PageBreakBottoms != null && HtmlContainer != null)
-                {
-                    // PageBreakBottoms is keyed by pagination slot, which is exactly what the fragment
-                    // records - no re-deriving the slot from a scroll offset.
-                    if (PageBreakBottoms.TryGetValue(fragment.FragmentainerIndex, out var pageBreakBottom))
-                    {
-                        var pageBreakBottomVisual = pageBreakBottom - fragment.OriginY;
-                        if (pageBreakBottomVisual < actualRect.Bottom)
-                        {
-                            rectForBorders = new RRect(
-                                actualRect.Left,
-                                actualRect.Top,
-                                actualRect.Width,
-                                pageBreakBottomVisual - actualRect.Top);
-                        }
-                    }
-                }
-
-                BordersDrawHandler.DrawBoxBorders(g, this, rectForBorders, i == 0, i == lines.Count - 1);
-            }
-
-            if (ColumnRuleSegments is { Count: > 0 } && ActualColumnRuleWidth > 0)
-            {
-                PaintColumnRules(g, fragment.OriginY, clip);
-            }
-
-            PaintWords(g, fragment);
-
-            for (var i = 0; i < lines.Count; i++)
-            {
-                var actualRect = lines[i].Rect;
-
-                if (IsRectVisible(actualRect, clip))
-                {
-                    PaintDecoration(g, actualRect, i == 0, i == lines.Count - 1, GetFirstLineStyleForRect(lines[i].Line));
-                }
-            }
-
-            var stackingContextBoxes = StackingOrder.Flatten(fragment);
-
-            foreach (var layerBoxes in StackingOrder.ByLayers(stackingContextBoxes))
-            {
-                // Split paint to handle z-order, per CSS2.1 Appendix E's within-a-stacking-context
-                // order: in-flow block-level descendants, then non-positioned floats, then in-flow
-                // inline-level descendants (text and inline replaced content), then positioned
-                // descendants. Block and inline normal-flow content used to share a single pass here
-                // (painted in tree order with no float-relative ordering guarantee at all) - Acid2's own
-                // ".eyes" trap (a block, a float, and an inline replaced <object> as siblings, each
-                // required to paint in a different layer) depends on inline being its own later pass.
-                //
-                // A plain (non-inline-itself) box whose entire content is inline - e.g. Acid2's own
-                // "#eyes-a" div, a block wrapper around nothing but its resolved inline <object> image -
-                // is treated as belonging to the inline pass too, via StackingOrder.ActsAsInline. Its
-                // own recursive Paint() call is what actually paints its inline child (through ITS OWN
-                // nested stacking loop), so deferring that whole call to this stacking context's inline
-                // pass is what makes the child paint after this context's float pass, matching Appendix
-                // E, without needing to hoist the descendant out of its normal DOM position/ancestor at
-                // all (unlike the out-of-flow float/absolute/fixed hoisting StackingOrder.Flatten
-                // already does - that mechanism moves a box's paint call across ancestor boundaries
-                // entirely, which isn't needed or wanted here since "#eyes-a" itself already belongs to
-                // this stacking context's own direct children).
-                foreach (var p in layerBoxes)
-                {
-                    if (!StackingOrder.ActsAsInline(p.Box) && p.Box.Position != CssConstants.Absolute && p.Box is { IsFixed: false, IsFloated: false })
-                        await PaintStackingParticipant(g, p);
-                }
-
-                foreach (var p in layerBoxes)
-                {
-                    if (p.Box.IsFloated)
-                        await PaintStackingParticipant(g, p);
-                }
-
-                foreach (var p in layerBoxes)
-                {
-                    if (StackingOrder.ActsAsInline(p.Box) && p.Box.Position != CssConstants.Absolute && p.Box is { IsFixed: false, IsFloated: false })
-                        await PaintStackingParticipant(g, p);
-                }
-
-                foreach (var p in layerBoxes)
-                {
-                    if (p.Box.Position == CssConstants.Absolute)
-                        await PaintStackingParticipant(g, p);
-                }
-
-                foreach (var p in layerBoxes)
-                {
-                    if (p.Box.IsFixed)
-                        await PaintStackingParticipant(g, p);
-                }
-            }
-
-            if (clipped)
-                g.PopClip();
-
-            if (paintMarkers)
-            {
-                var markerFragment = FindMarkerFragment(fragment);
-                if (markerFragment != null)
-                {
-                    await markerFragment.Box.PaintImpCore(g, markerFragment);
-                }
-            }
-
-            PaintContentImage(g, fragment);
-        }
-
-        /// <summary>
-        /// Paints one stacking-context participant discovered by <see cref="StackingOrder.Flatten"/>.
-        /// A hoisted participant (non-empty <see cref="StackingOrder.StackingParticipant.ClipAncestors"/>)
-        /// paints via this box's own paint loop rather than the ordinary nested parent-to-child
-        /// <see cref="Paint"/> cascade, so its ancestors' own <c>overflow: hidden</c> clipping - normally
-        /// picked up "for free" from still being active on the graphics clip stack during natural nested
-        /// painting - is never applied on its own. Re-apply it explicitly here instead, scoped to exactly
-        /// this participant's own paint call. A no-op for a direct plain child (empty ClipAncestors).
-        /// </summary>
-        private static async ValueTask PaintStackingParticipant(RGraphics g, StackingOrder.StackingParticipant participant)
-        {
-            var fragment = participant.Fragment;
-            var pushedClips = RenderUtils.PushAncestorOverflowClips(g, participant.ClipAncestors, fragment.OriginY);
-
-            await fragment.Box.Paint(g, fragment);
-
-            for (var i = 0; i < pushedClips; i++)
-                g.PopClip();
-        }
-
-        /// <summary>
-        /// Draws the vertical rule lines between columns of a multi-column container, one segment per
-        /// gap per page-row (see <see cref="ColumnRuleSegments"/>).
-        /// </summary>
-        private void PaintColumnRules(RGraphics g, double originY, RRect clip)
-        {
-            var pen = g.GetPen(ActualColumnRuleColor);
-            pen.Width = ActualColumnRuleWidth;
-            pen.DashStyle = ColumnRuleStyle switch
-            {
-                CssConstants.Dashed => RDashStyle.Dash,
-                CssConstants.Dotted => RDashStyle.Dot,
-                _ => RDashStyle.Solid,
-            };
-
-            // Column rules are recorded by the columns engine in document space, so they need the same
-            // origin shift the builder already applied to every other rectangle on this page.
-            foreach (var (x, top, bottom) in ColumnRuleSegments!)
-            {
-                var visualX = x;
-                var visualTop = top - originY;
-                var visualBottom = bottom - originY;
-
-                if (!IsRectVisible(new RRect(visualX - 1, visualTop, 2, visualBottom - visualTop), clip)) continue;
-
-                g.DrawLine(pen, visualX, visualTop, visualX, visualBottom);
-            }
-        }
-
-        private static bool IsRectVisible(RRect rect, RRect clip)
-        {
-            rect.X -= 2;
-            rect.Width += 2;
-            clip.Intersect(rect);
-
-            // See the zero-area-intersection fix/comment on CssBox.Paint's visibility cull
-            // (GitHub issue #113) - a rect merely touching the clip edge isn't actually visible.
-            return clip.Width > VisibilityClipEpsilon && clip.Height > VisibilityClipEpsilon;
-        }
-
-        /// <summary>
-        /// Resolves the <c>::first-line</c> style (if any) that applies to a rectangle painted for
-        /// <paramref name="lineBox"/> - true exactly when the line's own owner (the block establishing
-        /// that inline formatting context - see <see cref="CssLineBox.OwnerBox"/>) has a resolved
-        /// first-line style and this is genuinely that owner's first line. Used to make
-        /// <see cref="PaintBackground"/>/<see cref="PaintDecoration"/> first-line-aware without either
-        /// method needing to know its own relationship to the block establishing its inline formatting
-        /// context - the line box already knows.
-        /// </summary>
-        private static CssBox? GetFirstLineStyleForRect(CssLineBox? lineBox) =>
-            lineBox is not null && lineBox.OwnerBox.ResolvedFirstLineStyle is not null && lineBox == lineBox.OwnerBox.LineBoxes.FirstOrDefault()
-                ? lineBox.OwnerBox.ResolvedFirstLineStyle
-                : null;
-
-        /// <summary>
-        /// Paints the background of the box
-        /// </summary>
-        /// <param name="g">the device to draw into</param>
-        /// <param name="rect">the bounding rectangle to draw in</param>
-        /// <param name="isFirst">is it the first rectangle of the element</param>
-        /// <param name="firstLineStyle">
-        /// When set, this rect is on the target's first formatted line under a <c>::first-line</c>
-        /// rule - its resolved <c>background-color</c> is used instead of this box's own. Only
-        /// <c>background-color</c> is first-line-aware, not <c>background-image</c>/-position/-size/
-        /// -repeat/-origin/-clip layers (a documented narrowing - see docs/html-css-support.md);
-        /// those, like border/padding/border-radius (genuine box-model properties CSS2.1 never allows
-        /// on <c>::first-line</c> at all), always come from this box's own resolved style.
-        /// </param>
-        protected void PaintBackground(RGraphics g, RRect rect, bool isFirst, CssBox? firstLineStyle = null)
-        {
-            if (rect is { Width: > 0, Height: > 0 })
-            {
-                RRect BoxModelRect(string value) => value switch
-                {
-                    CssConstants.BorderBox => rect,
-                    CssConstants.ContentBox => new RRect(
-                        rect.X + ActualBorderLeftWidth + ActualPaddingLeft,
-                        rect.Y + ActualBorderTopWidth  + ActualPaddingTop,
-                        rect.Width  - ActualBorderLeftWidth - ActualBorderRightWidth  - ActualPaddingLeft - ActualPaddingRight,
-                        rect.Height - ActualBorderTopWidth  - ActualBorderBottomWidth - ActualPaddingTop  - ActualPaddingBottom),
-                    _ => new RRect(
-                        rect.X + ActualBorderLeftWidth,
-                        rect.Y + ActualBorderTopWidth,
-                        rect.Width  - ActualBorderLeftWidth - ActualBorderRightWidth,
-                        rect.Height - ActualBorderTopWidth  - ActualBorderBottomWidth),
-                };
-
-                // background-origin/background-clip are themselves comma-list (per-layer) properties,
-                // just like background-image/-position/-size - resolved per layer inside the loop below,
-                // not once for the whole box.
-                var originLayers = BackgroundLayerResolver.SplitLayers(BackgroundOrigin);
-                var clipLayers   = BackgroundLayerResolver.SplitLayers(BackgroundClip);
-                var viewportRect = HtmlContainer!.PageBoxRect;
-
-                var actualBackgroundColor = firstLineStyle?.ActualBackgroundColor ?? ActualBackgroundColor;
-                RBrush? solidBrush = RenderUtils.IsColorVisible(actualBackgroundColor)
-                    ? g.GetSolidBrush(actualBackgroundColor)
-                    : null;
-
-                // background-color is always the bottom-most layer (CSS Backgrounds 3 §3.6/§3.8) -
-                // painted BEFORE any background-image/gradient layer so those layers appear on top of
-                // it, not hidden beneath an opaque color.
-                if (solidBrush != null)
-                {
-                    // background-color is not itself a layered property: when background-clip has
-                    // multiple values, the solid fill uses the LAST (bottom-most) one, independent of
-                    // how many background-image layers exist - including zero.
-                    var colorClipRect = BoxModelRect(clipLayers[^1]);
-
-                    RGraphicsPath? colorRoundedClipPath = null;
-                    if (IsRounded)
-                    {
-                        var radii = ComputeRadii(colorClipRect);
-                        colorRoundedClipPath = RenderUtils.GetRoundRect(g, colorClipRect,
-                            radii.TLX, radii.TLY, radii.TRX, radii.TRY,
-                            radii.BRX, radii.BRY, radii.BLX, radii.BLY);
-                    }
-
-                    PaintClippedBrush(g, solidBrush, colorClipRect, colorRoundedClipPath);
-                    colorRoundedClipPath?.Dispose();
-                }
-
-                // Then paint image/gradient layers back-to-front (last in the comma-list = bottom-most
-                // image layer, but still on top of the solid color) so the first-declared layer ends up
-                // visually on top.
-                var layersToPaint = BackgroundImages != null
-                    ? Enumerable.Range(0, BackgroundImages.Count).Reverse()
-                    : Enumerable.Empty<int>();
-
-                foreach (int layerIndex in layersToPaint)
-                {
-                    var originRect = BoxModelRect(BackgroundLayerResolver.LayerAt(originLayers, layerIndex));
-                    var clipRect   = BoxModelRect(BackgroundLayerResolver.LayerAt(clipLayers, layerIndex));
-
-                    RGraphicsPath? roundedClipPath = null;
-                    if (IsRounded)
-                    {
-                        var radii = ComputeRadii(clipRect);
-                        roundedClipPath = RenderUtils.GetRoundRect(g, clipRect,
-                            radii.TLX, radii.TLY, radii.TRX, radii.TRY,
-                            radii.BRX, radii.BRY, radii.BLX, radii.BLY);
-                    }
-
-                    void DrawBrush(RBrush brush) => PaintClippedBrush(g, brush, clipRect, roundedClipPath);
-
-                    CssImagePainter.Paint(g, BackgroundImages![layerIndex], layerIndex, isFirst, originRect, clipRect,
-                        roundedClipPath, BackgroundPosition, BackgroundSize, BackgroundRepeat, BackgroundAttachment,
-                        viewportRect, this, DrawBrush);
-
-                    roundedClipPath?.Dispose();
-                }
-            }
-        }
-
-        /// <summary>
-        /// Fills <paramref name="brush"/> clipped to <paramref name="roundedClipPath"/> (border-radius
-        /// case) or <paramref name="clipRect"/> (rectangular case), and disposes the brush afterward.
-        /// Shared by every per-layer background-image/gradient draw and the final solid-color fill.
-        /// </summary>
-        private void PaintClippedBrush(RGraphics g, RBrush brush, RRect clipRect, RGraphicsPath? roundedClipPath)
-        {
-            // TODO:a handle it correctly (tables background)
-            object? prevMode = null;
-            if (HtmlContainer is { AvoidGeometryAntialias: false } && IsRounded)
-                prevMode = g.SetAntiAliasSmoothingMode();
-
-            if (roundedClipPath != null)
-                g.DrawPath(brush, roundedClipPath);
-            else
-                g.DrawRectangle(brush, clipRect.X, clipRect.Y, clipRect.Width, clipRect.Height);
-
-            g.ReturnPreviousSmoothingMode(prevMode);
-            brush.Dispose();
-        }
-
-        /// <summary>
-        /// Paints this box's <c>box-shadow</c> layers of one kind (outset or inset) for a single border-box
-        /// fragment (CSS Backgrounds &amp; Borders Level 3 §5). PDF has no native blur, so a blurred shadow is
-        /// approximated with vector geometry: a stack of concentric, overlapping rounded-rect fills each at a
-        /// small constant alpha, so their source-over accumulation ramps the shadow color's alpha from zero at
-        /// the outer blur edge to the full color where every layer overlaps (the interior). Because the layers
-        /// overlap rather than abut, there are no partial-alpha shared edges to double-blend into seam lines,
-        /// and the corners round off automatically over the blur radius. Layers paint last-listed first so the
-        /// first-declared shadow ends up on top.
-        /// </summary>
-        private void PaintBoxShadows(RGraphics g, RRect borderBox, bool inset)
-        {
-            var layers = BoxShadowGrammar.TryParse(CssValueParser.GetCssTokens(BoxShadow));
-            if (layers is null || layers.Count == 0) return;
-
-            for (var i = layers.Count - 1; i >= 0; i--)
-            {
-                var layer = layers[i];
-                if (layer.Inset != inset) continue;
-
-                var dx = CssValueParser.ParseLength(layer.OffsetX, 0, this);
-                var dy = CssValueParser.ParseLength(layer.OffsetY, 0, this);
-                var blur = CssValueParser.ParseLength(layer.Blur, 0, this);
-                var spread = CssValueParser.ParseLength(layer.Spread, 0, this);
-                var color = ResolveShadowColor(layer.Color);
-
-                if (inset)
-                    PaintInsetShadow(g, borderBox, dx, dy, blur, spread, color);
-                else
-                    PaintOutsetShadow(g, borderBox, dx, dy, blur, spread, color);
-            }
-        }
-
-        /// <summary>Resolves a shadow layer's authored color string to an <see cref="RColor"/>; a null/omitted
-        /// or <c>currentColor</c> value uses the element's own text color (CSS Backgrounds 3 §5).</summary>
-        private RColor ResolveShadowColor(string? color) =>
-            string.IsNullOrEmpty(color) || color.Equals(CssConstants.CurrentColor, StringComparison.OrdinalIgnoreCase)
-                ? ActualColor
-                : GetActualColor(color);
-
-        private void PaintOutsetShadow(RGraphics g, RRect borderBox, double dx, double dy, double blur, double spread, RColor color)
-        {
-            // The shadow shape is the border box, translated by the offset and expanded by spread on all
-            // sides (a negative spread shrinks it).
-            var shadowRect = new RRect(
-                borderBox.X + dx - spread,
-                borderBox.Y + dy - spread,
-                borderBox.Width + 2 * spread,
-                borderBox.Height + 2 * spread);
-
-            if (shadowRect.Width <= 0 || shadowRect.Height <= 0) return;
-
-            // The shadow shape's own corner radii: the box's border radius grown by the spread (a sharp,
-            // zero-radius corner stays sharp).
-            var baseRadii = ShadowCornerRadii(borderBox, spread);
-
-            if (blur <= 0)
-            {
-                var brush = g.GetSolidBrush(color);
-                if (baseRadii.IsRounded)
-                {
-                    var path = BuildLayerRoundRect(g, shadowRect, baseRadii, 0);
-                    g.DrawPath(brush, path);
-                    path.Dispose();
-                }
-                else
-                {
-                    g.DrawRectangle(brush, shadowRect.X, shadowRect.Y, shadowRect.Width, shadowRect.Height);
-                }
-                brush.Dispose();
-                return;
-            }
-
-            var steps = BlurSteps(blur);
-            var layerColors = ComputeBlurLayerColors(color, steps);
-            if (layerColors.Length == 0) return;
-
-            // Draw outermost (largest) to innermost. A point at signed outward distance x is covered by
-            // layers 0..K(x); painting them in this order accumulates to the target alpha ramp for that point.
-            for (var k = 0; k < steps; k++)
-            {
-                var d = blur - 2 * blur * k / (steps - 1); // +blur (outer) .. -blur (inner)
-                var rect = new RRect(shadowRect.X - d, shadowRect.Y - d, shadowRect.Width + 2 * d, shadowRect.Height + 2 * d);
-                if (rect.Width <= 0 || rect.Height <= 0) continue;
-
-                var brush = g.GetSolidBrush(layerColors[k]);
-                var path = BuildLayerRoundRect(g, rect, baseRadii, d);
-                g.DrawPath(brush, path);
-                path.Dispose();
-                brush.Dispose();
-            }
-        }
-
-        private void PaintInsetShadow(RGraphics g, RRect borderBox, double dx, double dy, double blur, double spread, RColor color)
-        {
-            var paddingBox = new RRect(
-                borderBox.X + ActualBorderLeftWidth,
-                borderBox.Y + ActualBorderTopWidth,
-                borderBox.Width - ActualBorderLeftWidth - ActualBorderRightWidth,
-                borderBox.Height - ActualBorderTopWidth - ActualBorderBottomWidth);
-
-            if (paddingBox.Width <= 0 || paddingBox.Height <= 0) return;
-
-            // The lit "hole" = the padding box, translated by the offset and shrunk by spread. The shadow is
-            // the inverse (the region between the padding box and this inner shape), clipped to the padding box.
-            var inner = new RRect(
-                paddingBox.X + dx + spread,
-                paddingBox.Y + dy + spread,
-                Math.Max(0, paddingBox.Width - 2 * spread),
-                Math.Max(0, paddingBox.Height - 2 * spread));
-
-            RGraphicsPath? clipPath = null;
-            if (IsRounded)
-            {
-                clipPath = BuildLayerRoundRect(g, paddingBox, ShadowCornerRadii(borderBox, spread: 0), 0);
-                g.PushClip(clipPath);
-            }
-            else
-            {
-                g.PushClip(paddingBox);
-            }
-
-            if (blur <= 0)
-            {
-                // Solid ring = padding box minus the inner hole.
-                FillRingRects(g, paddingBox, inner, color);
-            }
-            else
-            {
-                var steps = BlurSteps(blur);
-                var layerColors = ComputeBlurLayerColors(color, steps);
-
-                // Each layer is the padding box with a rectangular hole punched out (via an even-odd fill,
-                // since the PDF backend has no clip-subtract primitive). The hole grows from
-                // inner-deflated-by-blur (drawn first) to inner-inflated-by-blur, so a point out toward the
-                // padding edge is covered by every layer (full color) and one deep in the hole by none
-                // (transparent) - the inset falloff, fading inward over the blur radius. The already-pushed
-                // padding-box clip trims each ring to the (possibly rounded) padding box.
-                for (var k = 0; k < layerColors.Length; k++)
-                {
-                    var d = -blur + 2 * blur * k / (steps - 1);
-                    var hole = new RRect(inner.X - d, inner.Y - d, inner.Width + 2 * d, inner.Height + 2 * d);
-                    var brush = g.GetSolidBrush(layerColors[k]);
-                    var ring = BuildRingPath(g, paddingBox, hole);
-                    g.DrawPath(brush, ring);
-                    ring.Dispose();
-                    brush.Dispose();
-                }
-            }
-
-            g.PopClip();
-            clipPath?.Dispose();
-        }
-
-        /// <summary>The shadow shape's per-corner radii: the box's <c>border-radius</c> grown by
-        /// <paramref name="spread"/> where non-zero, with sharp corners staying sharp.</summary>
-        private BorderRadii ShadowCornerRadii(RRect borderBox, double spread)
-        {
-            var r = ComputeRadii(borderBox);
-            double Adj(double v) => v > 0 ? Math.Max(0, v + spread) : 0;
-            return new BorderRadii(Adj(r.TLX), Adj(r.TLY), Adj(r.TRX), Adj(r.TRY),
-                                   Adj(r.BRX), Adj(r.BRY), Adj(r.BLX), Adj(r.BLY));
-        }
-
-        /// <summary>Number of concentric fills used to approximate a blur of the given radius (in points).</summary>
-        private static int BlurSteps(double blur) => Math.Clamp((int)Math.Round(blur * 2), 6, 40);
-
-        /// <summary>
-        /// Per-layer colors for the <paramref name="steps"/> concentric blur fills (index 0 = outermost,
-        /// faintest). The alpha of each layer is chosen so that the running source-over accumulation of
-        /// layers 0..k reaches a linear target ramp - <c>alpha[k] = 1 - (1 - T_k)/(1 - T_{k-1})</c> with
-        /// <c>T_k = Amax·(k+1)/steps</c> - so the interior reaches the shadow color's <b>full</b> alpha even
-        /// when it is fully opaque (a constant per-layer alpha would collapse an opaque color to a hard edge).
-        /// Empty (paints nothing) for a fully-transparent shadow color.
-        /// </summary>
-        private static RColor[] ComputeBlurLayerColors(RColor color, int steps)
-        {
-            var amax = color.A / 255.0;
-            if (amax <= 0) return [];
-
-            var colors = new RColor[steps];
-            var prevTarget = 0.0;
-
-            for (var k = 0; k < steps; k++)
-            {
-                var target = amax * (k + 1) / steps;
-                var a = 1 - (1 - target) / (1 - prevTarget);
-                prevTarget = target;
-
-                var alpha = Math.Clamp((int)Math.Round(a * 255), 1, 255);
-                colors[k] = RColor.FromArgb(alpha, color.R, color.G, color.B);
-            }
-
-            return colors;
-        }
-
-        /// <summary>
-        /// Builds a rounded-rect path for one concentric shadow layer at signed outward distance
-        /// <paramref name="d"/> from the shadow's sharp edge. Each corner's radius is the base (sharp-shape)
-        /// radius plus <paramref name="d"/>, clamped non-negative - so outer layers of even a square box round
-        /// off over the blur radius, matching how a real blurred shadow's corners soften.
-        /// </summary>
-        private static RGraphicsPath BuildLayerRoundRect(RGraphics g, RRect rect, BorderRadii baseRadii, double d)
-        {
-            double R(double b) => Math.Max(0, b + d);
-            return RenderUtils.GetRoundRect(g, rect,
-                R(baseRadii.TLX), R(baseRadii.TLY), R(baseRadii.TRX), R(baseRadii.TRY),
-                R(baseRadii.BRX), R(baseRadii.BRY), R(baseRadii.BLX), R(baseRadii.BLY));
-        }
-
-        /// <summary>Fills the region of <paramref name="outer"/> that lies outside <paramref name="hole"/>
-        /// with a solid color, as four axis-aligned rectangles (the hole is clamped to the outer bounds).</summary>
-        private static void FillRingRects(RGraphics g, RRect outer, RRect hole, RColor color)
-        {
-            var hl = Math.Max(outer.Left, hole.Left);
-            var ht = Math.Max(outer.Top, hole.Top);
-            var hr = Math.Min(outer.Right, hole.Right);
-            var hb = Math.Min(outer.Bottom, hole.Bottom);
-            if (hr < hl) hr = hl;
-            if (hb < ht) hb = ht;
-
-            var brush = g.GetSolidBrush(color);
-            FillRect(g, brush, outer.Left, outer.Top, outer.Width, ht - outer.Top);          // top
-            FillRect(g, brush, outer.Left, hb, outer.Width, outer.Bottom - hb);              // bottom
-            FillRect(g, brush, outer.Left, ht, hl - outer.Left, hb - ht);                    // left
-            FillRect(g, brush, hr, ht, outer.Right - hr, hb - ht);                           // right
-            brush.Dispose();
-        }
-
-        private static void FillRect(RGraphics g, RBrush brush, double x, double y, double width, double height)
-        {
-            if (width > 0 && height > 0)
-                g.DrawRectangle(brush, x, y, width, height);
-        }
-
-        /// <summary>Builds an even-odd fill path of <paramref name="outer"/> with a rectangular
-        /// <paramref name="hole"/> punched out - a filled ring. Used for inset-shadow falloff layers, since the
-        /// PDF backend has no clip-subtract primitive.</summary>
-        private static RGraphicsPath BuildRingPath(RGraphics g, RRect outer, RRect hole)
-        {
-            var path = g.GetGraphicsPath();
-
-            path.Start(outer.Left, outer.Top);
-            path.LineTo(outer.Right, outer.Top);
-            path.LineTo(outer.Right, outer.Bottom);
-            path.LineTo(outer.Left, outer.Bottom);
-            path.CloseFigure();
-
-            path.AddMove(hole.Left, hole.Top);
-            path.LineTo(hole.Right, hole.Top);
-            path.LineTo(hole.Right, hole.Bottom);
-            path.LineTo(hole.Left, hole.Bottom);
-            path.CloseFigure();
-
-            path.FillMode = RFillMode.EvenOdd;
-            return path;
-        }
-
-        /// <summary>
-        /// Paint all the words in the box.
-        /// </summary>
-        /// <param name="g">the device to draw into</param>
-        /// <param name="fragment">the fragment whose words are painted</param>
-        private void PaintWords(RGraphics g, BoxFragment fragment)
-        {
-            if (Width is null or { Length: <= 0 }) return;
-
-            var isRtl = Direction == CssConstants.Rtl;
-
-            foreach (var wordFragment in fragment.Words)
-            {
-                var word = wordFragment.Word;
-                if (word.IsLineBreak || word.IsImage) continue;
-                var clip = g.GetClip();
-                clip.Intersect(wordFragment.Rect);
-
-                // A word whose box was relocated to the next page's content top (keep-with-next,
-                // break-inside:avoid, orphans/widows) sits exactly flush against the previous page's
-                // clip bottom - RRect.Intersect can land a hair off exact zero in either direction
-                // (floating-point rounding across the several arithmetic steps a relocated box's Y goes
-                // through), so neither RRect.Empty nor a strict zero check reliably catches it; the
-                // epsilon does. Without this, a fully-clipped (invisible on screen, but present in the
-                // content stream and text-extraction layer) duplicate of the word painted on the page it
-                // just left. See GitHub issue #113.
-                if (clip.Width <= VisibilityClipEpsilon || clip.Height <= VisibilityClipEpsilon) continue;
-
-                // A word on the target's first formatted line, under a ::first-line rule, uses that
-                // resolved shadow box's font/color/letter-spacing instead of this box's own - it was
-                // already measured against this same styleSource (see ApplyFirstLineStyleOverride), so
-                // word.Top/Height are already consistent with it.
-                var styleSource = word.FirstLineStyle ?? this;
-
-                // A fragment drawn with a different font than the box's own ActualFont - a synthesized
-                // small-caps run (smaller size) or a per-codepoint fallback face (different metrics) - is
-                // top-anchored at the same word.Top (the shared line box's top), so without correction its
-                // baseline would sit at a different height than its full-size neighbors'. Shift down by the
-                // ascent difference so every fragment's baseline lines up. This is exactly 0 for an ordinary
-                // word (font == ActualFont), so it is a no-op there.
-                var font = ResolveWordFont(word, styleSource);
-                var baselineAdjust = styleSource.ActualFont.Ascent - font.Ascent;
-                var wordPoint = new RPoint(wordFragment.Rect.X, wordFragment.Rect.Y + baselineAdjust);
-                var text = word.FirstLineText ?? word.Text!;
-                g.DrawString(text, font, styleSource.ActualColor, wordPoint, new RSize(word.Width, word.Height), isRtl, styleSource.ActualLetterSpacing, styleSource.ActualFontPalette);
-            }
-        }
-
-        /// <summary>
-        /// Paints the text decoration (underline/strike-through/over-line)
-        /// </summary>
-        /// <param name="g">the device to draw into</param>
-        /// <param name="rectangle"> </param>
-        /// <param name="isFirst"> </param>
-        /// <param name="isLast"> </param>
-        /// <param name="firstLineStyle">
-        /// When set, this rectangle is on the target's first formatted line under a <c>::first-line</c>
-        /// rule - its resolved text-decoration/color/font (for underline-offset) are used instead of
-        /// this box's own.
-        /// </param>
-        protected void PaintDecoration(RGraphics g, RRect rectangle, bool isFirst, bool isLast, CssBox? firstLineStyle = null)
-        {
-            // The `text-decoration` shorthand is expanded into these longhands by the CSS-OM (Layer A) before
-            // it ever reaches the box, so the painter reads the longhands directly. text-decoration-line may
-            // carry several space-separated line keywords (e.g. "underline overline"); each is drawn.
-            var styleSource = firstLineStyle ?? this;
-            var textDecorationLine = styleSource.TextDecorationLine;
-            var textDecorationStyle = styleSource.TextDecorationStyle;
-            var textDecorationColor = styleSource.TextDecorationColor;
-
-            if (string.IsNullOrEmpty(textDecorationLine) || textDecorationLine == CssConstants.None)
-                return;
-
-            if (string.IsNullOrEmpty(textDecorationStyle))
-            {
-                textDecorationStyle = CssConstants.Solid;
-            }
-
-            if (!string.IsNullOrEmpty(textDecorationColor) && !HtmlContainer!.CssParser.IsColorValid(textDecorationColor))
-            {
-                textDecorationColor = string.Empty;
-            }
-
-            var textDecorationActualColor = string.IsNullOrEmpty(textDecorationColor) ? styleSource.ActualColor : HtmlContainer!.CssParser.ParseColor(textDecorationColor);
-
-            double x1 = rectangle.X;
-            if (isFirst)
-                x1 += ActualPaddingLeft + ActualBorderLeftWidth;
-
-            double x2 = rectangle.Right;
-            if (isLast)
-                x2 -= ActualPaddingRight + ActualBorderRightWidth;
-
-            var dashStyle = textDecorationStyle switch
-            {
-                CssConstants.Solid => RDashStyle.Solid,
-                CssConstants.Double => RDashStyle.Solid,
-                CssConstants.Dotted => RDashStyle.Dot,
-                CssConstants.Dashed => RDashStyle.Dash,
-                CssConstants.Wavy => RDashStyle.Solid,
-                _ => RDashStyle.Solid
-            };
-
-            var pen = g.GetPen(textDecorationActualColor);
-            pen.Width = 1;
-            pen.DashStyle = dashStyle;
-
-            // text-decoration-line may list several keywords (e.g. "underline overline"); draw each.
-            var bottomInset = ActualPaddingBottom - ActualBorderBottomWidth;
-            foreach (var line in textDecorationLine.Split(' ', StringSplitOptions.RemoveEmptyEntries))
-            {
-                double y = line switch
-                {
-                    CssConstants.Underline => Math.Round(rectangle.Top + styleSource.ActualFont.UnderlineOffset),
-                    CssConstants.LineThrough => rectangle.Top + rectangle.Height / 2f,
-                    CssConstants.Overline => rectangle.Top,
-                    _ => double.NaN
-                };
-
-                if (double.IsNaN(y)) continue;
-
-                y -= bottomInset;
-                g.DrawLine(pen, x1, y, x2, y);
-            }
         }
 
         /// <summary>
