+++ src/PeachPDF/Html/Core/Dom/CssBoxObject.cs
@@ -5,7 +5,6 @@ using PeachPDF.Html.Core.Utils;
 using PeachPDF.Svg;
 using System;
 using System.Threading.Tasks;
-using PeachPDF.Html.Core.Fragments;
 
 namespace PeachPDF.Html.Core.Dom
 {
@@ -64,36 +63,22 @@ namespace PeachPDF.Html.Core.Dom
             }
         }
 
-        protected override async ValueTask PaintImpCore(RGraphics g, BoxFragment fragment)
-        {
-            await EnsureResolved();
+        /// <summary>
+        /// Whether the <c>data</c> resource resolved to a supported image, making this box a replaced
+        /// element. Only meaningful once <see cref="EnsureResolved"/> has run.
+        /// </summary>
+        internal bool IsReplaced => _isReplaced;
 
-            if (!_isReplaced)
-            {
-                await base.PaintImpCore(g, fragment);
-                return;
-            }
+        /// <summary>
+        /// The phantom word carrying this box's replaced content, whose own rectangle positions the
+        /// image within the box. Null until (and unless) resolution succeeds.
+        /// </summary>
+        internal CssRectImage? ReplacedWord => _imageWord;
 
-            var rect = fragment.PrimaryRect;
-
-            var clipped = RenderUtils.ClipGraphicsByOverflow(g, this, fragment.OriginY);
-
-            PaintBackground(g, rect, true);
-            BordersDrawHandler.DrawBoxBorders(g, this, rect, true, true);
-
-            if (!fragment.TryGetWordRect(_imageWord!, out var r))
-                r = rect;
-            r.Height -= ActualBorderTopWidth + ActualBorderBottomWidth + ActualPaddingTop + ActualPaddingBottom;
-            r.Y += ActualBorderTopWidth + ActualPaddingTop;
-            r.X = Math.Floor(r.X);
-            r.Y = Math.Floor(r.Y);
-
-            // object-fit / object-position honored via the shared replaced-content renderer.
-            ReplacedContentRenderer.Paint(g, r, _imageWord!.Image, _svgDocument, this);
-
-            if (clipped)
-                g.PopClip();
-        }
+        /// <summary>
+        /// The parsed SVG scene graph when the resolved resource was an SVG image, else null.
+        /// </summary>
+        internal SvgDocument? SvgDocument => _svgDocument;
 
         /// <summary>
         /// Resolves whether this object is replaced content (a supported image) or must fall back to
@@ -103,7 +88,7 @@ namespace PeachPDF.Html.Core.Dom
         /// resolution deterministic and independent of network reachability, which matters for the
         /// real Acid2 fixture's intentionally-unreachable middle <c>&lt;object&gt;</c>.
         /// </summary>
-        private async ValueTask EnsureResolved()
+        internal async ValueTask EnsureResolved()
         {
             if (_resolved)
                 return;
